chore: move making default config to shared utils and fix default configs
