+++ packages/ui/src/ui/store/selectors/dashboard2/dashboard.ts
@@ -1,8 +1,4 @@
 import {createSelector} from '@reduxjs/toolkit';
-import {DashKitProps} from '@gravity-ui/dashkit';
-
-import find_ from 'lodash/find';
-import remove_ from 'lodash/remove';
 
 import {RootState} from '../../../store/reducers';
 import {accountsApi} from '../../../store/api/accounts';
@@ -10,6 +6,8 @@ import {getSettingsData} from '../../../store/selectors/settings/settings-base';
 import {getCluster, getCurrentUserName} from '../../../store/selectors/global';
 
 import {dashboardConfig} from '../../../constants/dashboard2';
+
+import {makeDefaultConfig} from '../../../utils/dashboard2/make-default-config';
 
 export const getUsableAccountsResult = (state: RootState) => {
     const selector = accountsApi.endpoints.usableAccounts.select(undefined);
@@ -27,27 +25,7 @@ export const getDashboardConfig = createSelector(
             return data[`local::${cluster}::dashboard::config`];
         }
 
-        const prevItems = [...dashboardConfig.items.slice(0, 4)];
-
-        // we always have queries item in default config(its a constant)
-        const queriesItem = find_(prevItems, (item) => item.type === 'queries')!;
-        // new item cause of default username
-        const newItem = {
-            ...queriesItem,
-            data: {
-                ...queriesItem?.data,
-                authors: [{value: username, type: 'users'}],
-            },
-        };
-
-        remove_(prevItems, (item) => item.type === 'queries');
-
-        const config: DashKitProps['config'] = {
-            ...dashboardConfig,
-            items: [...prevItems, newItem],
-        };
-
-        return config;
+        return makeDefaultConfig(username);
     },
 );
